refactor(mctrl): rename module segments into factory for consistency
