+++ src/mission_control/src/factory/store.rs
